% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_FBasicFunctions.R
\name{R_F77}
\alias{R_F77}
\title{Add 77 to a real number using a Fortran77 function}
\usage{
R_F77(X)
}
\arguments{
\item{X}{A real number}
}
\value{
A real number
}
\description{
Add 77 to a real number using a Fortran77 function
}
\examples{
x = 1
R_F77(x)
}
